uint8_t_arg

+++ LEDPot_Task/main/library/wr_LED.h
@@ -12,7 +12,7 @@
 void on_led(void);
 void off_led(void);
 void configure_led(void);
-void chngLEDcolr(int *pVColor);
+void chngLEDcolr(uint8_t *pVColor);
 
 /*
 !DONT define variables with static and Done define static functions (their visiblity is to particular translation unit)
